Adding a cast to get rid of a compiler warning
git-svn-id: http://svn.code.sf.net/p/nitro-nitf/code/trunk@1048 7dc3ea37-dbd3-4c13-bb9a-7ccd57e75167

+++ c/nitf/source/Writer.c
@@ -303,7 +303,7 @@ NITFPRIV(NITF_BOOL) writeInt64Field(nitf_Writer * writer,
     char buf[20];
 
     memset(buf, '\0', 20);
-    NITF_SNPRINTF(buf, 20, "%lld", field);
+    NITF_SNPRINTF(buf, 20, "%lld", (long long int)field);
 
     if (padString(writer, buf, length, fill, fillDir, error))
     {
